Extract InputTab to separate component
